feat(gnl_file): add basic library for gnl file struct
I didn't make any new functions for it or anything, it just has the
simple stuff I made for get_next_line. All the struct is, is a file
descriptor variable and a buffer string. It's meant to be used in
get_next_line.c, and not really as an actual libft thing (hence the
separation into the header file).