refactor(dispatch table): index by character instead of number
This changes get_conversion to set the character to 0 on error, so I don't need to return an index
for error condition. I have also changed the jump table to be indexed by character now, and it's
hidden behind two layers of static functions, so I think it is nicely secured. I also add make_list
to the ft_printf_parse header, because it is the main parse function.