fix(makefile): indent rules, only run things once
This adds script-style if statements that check if there's anything to
clean before it gets cleaned, that way I don't print a cleaning line if
I don't need to (like when it gets run multiple times). You have to
have a semicolon to separate these bash statements, and they also don't
play nice with lines that don't end in a space (before the backslash).
This commit fixes all the little rule-scripts that I have for color
printing. It also makes sure that everything is the same style.