chore(strchr): clean ups and new functions
Adding: ft_strrchr_beg and ft_strrchr_i_beg These functions return the
last occurance of the character requested, or the beginning of the
string if the character isn't found. ft_strrchr_i_beg returns an index
of the character, instead of the pointer to the character.

This also touches up my strchr functions, they now take a character
parameter instead of an int, and I've cleaned up some of the logic (for
example, I've changed things like while (s && *s) { etc }, to if (!s) {
return } while (*s) { etc }).