feat(termcaps): functions to assist with termcaps
Add setup function which loads a terminal description, as well as a set
of functions to get strings and numbers, and a few helpers for specific
things, like moving the cursor. Eventually I'd like to make a function
or something that can load all the termcaps a program will need and
store them in a table somewhere, and then reference them later much
faster or something.