update(tc_setup): remove isatty check
Also take out old comment with things I don't need
I don't need to worry about freeing the tgetent call, because these
functions work with the terminfo database now, and that version of
tgetent doesn't work with the buffer pointer. I'm not sure if I'll
need to worry about it's memory anymore, and I'm still not sure of
what these functions do on the macs.