fix(term.c): free memory, reorganize
I now am freeing the terminal struct immediately if tcgetattr fails.
This makes sense, because it can't go anywhere so it should clean up
after itself.

Move the function to store the new allocated terminal in init_terms to
the bottom, that way I can free the struct whenever on the inside of the
function if something goes wrong and don't have to worry about resetting
the holder.

Change the hold function to take a double pointer. This lets me reset
the pointer back to 0 (by sending the address of the struct after I
delete it).