add(iter_pop): take dnode from head of iter
Add an ft_iter_pop_head function to return the dnode at the head of the
iter. Switch the rm_head function to use this. This function will make
the ft_iter_del easier, because I can just yoink a node from the top and
then free it with a del function. I'll probably end up coming up with a
better/more consistent structure for all these sub-libraries soon, it's
starting to bother me how things can be so different.